codecrafters submit [skip ci]

+++ src/main/java/Main.java
@@ -175,8 +175,8 @@ public class Main {
                                 else if("BLPOP".equals(inp[0]))
                                 {
                                     String key = inp[1];
-                                    int timeout=Integer.parseInt(inp[2]);
-                                    long endTime = System.currentTimeMillis() + timeout*1000;
+                                    float timeout=Float.parseFloat(inp[2]);
+                                    long endTime = System.currentTimeMillis() + (long)timeout*1000;
                                     List<String> list = listsMap.computeIfAbsent(key, k->new CopyOnWriteArrayList<>());
                                     String removed = null;
                                     List<String> removedArray = null;
